fix: reuse helper from ast-utils
