+++ packages/eslint-plugin-svelte/src/rules/derived-has-same-inputs-outputs.ts
@@ -3,34 +3,7 @@ import type { Variable } from '@typescript-eslint/scope-manager';
 import { createRule } from '../utils/index.js';
 import type { RuleContext, RuleFixer } from '../types.js';
 import { extractStoreReferences } from './reference-helpers/svelte-store.js';
-import { getScope } from '../utils/ast-utils.js';
-
-function findVariableForName(
-	context: RuleContext,
-	node: TSESTree.Node,
-	name: string,
-	expectedName: string
-): { hasConflict: boolean; variable: Variable | null } {
-	const scope = getScope(context, node);
-	let variable: Variable | null = null;
-
-	for (const ref of scope.references) {
-		if (ref.identifier.name === expectedName) {
-			return { hasConflict: true, variable: null };
-		}
-	}
-
-	for (const v of scope.variables) {
-		if (v.name === expectedName) {
-			return { hasConflict: true, variable: null };
-		}
-		if (v.name === name) {
-			variable = v;
-		}
-	}
-
-	return { hasConflict: false, variable };
-}
+import { findVariableForReplacement } from '../utils/ast-utils.js';
 
 function createFixer(node: TSESTree.Node, variable: Variable | null, name: string) {
 	return function* fix(fixer: RuleFixer) {
@@ -92,7 +65,7 @@ export default createRule('derived-has-same-inputs-outputs', {
 			if (fnParam.type !== 'Identifier') return;
 			const expectedName = `$${args.name}`;
 			if (expectedName !== fnParam.name) {
-				const { hasConflict, variable } = findVariableForName(
+				const { hasConflict, variable } = findVariableForReplacement(
 					context,
 					fn.body,
 					fnParam.name,
@@ -136,7 +109,7 @@ export default createRule('derived-has-same-inputs-outputs', {
 				if (element && element.type === 'Identifier' && argName) {
 					const expectedName = `$${argName}`;
 					if (expectedName !== element.name) {
-						const { hasConflict, variable } = findVariableForName(
+						const { hasConflict, variable } = findVariableForReplacement(
 							context,
 							fn.body,
 							element.name,
