+++ packages/eslint-plugin-svelte/tests/fixtures/rules/require-store-callbacks-use-set-param/invalid/test01-input.svelte
@@ -21,4 +21,11 @@
 	readable(false, () => {
 		const set = 303;
 	});
+
+	insideACallback(() => {
+		const set = 303;
+		readable(false, () => {
+			set;
+		});
+	});
 </script>
